some fixes

+++ routes/user-route.js
@@ -6,18 +6,16 @@ const { addDoc, collection, setDoc, getDocs, query, where, updateDoc, arrayUnion
 const multer = require('multer');
 const path = require('path');
 const jwt = require('jsonwebtoken');
-const admin = require("firebase-admin");
+const admin = require('firebase-admin');
 const authMiddleware = require('../common/auth-middleware');
 const uuid = require('uuid-v4');
-const serviceAccount = require("../doc-master-server-firebase-adminsdk-8sor4-7f05846648.json");
+const serviceAccount = require('../doc-master-server-firebase-adminsdk-8sor4-7f05846648.json');
 const { ref, getDownloadURL, getMetadata, deleteObject, uploadBytesResumable, listAll } = require('firebase/storage');
-const { storage2 } = require('../fireBase-Config');
-const axios = require('axios');
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
-  databaseURL: "https://doc-master-server-rtdb.firebaseio.com",
-  storageBucket: "doc-master-server.appspot.com"
+  databaseURL: 'https://doc-master-server-rtdb.firebaseio.com',
+  storageBucket: 'doc-master-server.appspot.com'
 });
 
 const bucket = admin.storage().bucket();
@@ -44,20 +42,20 @@ const refreshTokens = async (req, res) => {
   const refreshTokenOrig = authHeader && authHeader.split(' ')[1];
 
   if (!refreshTokenOrig) {
-    return res.status(401).send("Missing token");
+    return res.status(401).send('Missing token');
   }
 
   jwt.verify(refreshTokenOrig, process.env.REFRESH_TOKEN_SECRET, async (err, userInfo) => {
     if (err) {
-      return res.status(403).send("Invalid token");
+      return res.status(403).send('Invalid token');
     }
 
     try {
-      const userQuery = query(collection(db, "users"), where("_uid", "==", userInfo.uid));
+      const userQuery = query(collection(db, 'users'), where('_uid', '==', userInfo.uid));
       const querySnapshot = await getDocs(userQuery);
 
       if (querySnapshot.empty) {
-        return res.status(403).send("Invalid token");
+        return res.status(403).send('Invalid token');
       }
 
       const userDocRef = querySnapshot.docs[0].ref;
@@ -67,7 +65,7 @@ const refreshTokens = async (req, res) => {
         if (user.tokens) {
           await updateDoc(userDocRef, { tokens: [] });
         }
-        return res.status(403).send("Invalid token");
+        return res.status(403).send('Invalid token');
       }
 
       const { accessToken, refreshToken } = generateTokens(user._uid.toString());
@@ -89,13 +87,13 @@ const registerUser = async (req, res) => {
   try {
     const { full_name, email, password } = req.body;
     if (!email || !password) {
-      return res.status(400).send("Missing email or password");
+      return res.status(400).send('Missing email or password');
     }
 
     const docRef = await createUserWithEmailAndPassword(auth, email, password);
     const userObj = docRef.user;
 
-    await addDoc(collection(db, "users"), {
+    await addDoc(collection(db, 'users'), {
       full_name,
       email,
       _uid: userObj?.uid,
@@ -111,11 +109,11 @@ const registerUser = async (req, res) => {
 // Get user information
 const getUser = async (req, res) => {
   const user = req.body.user;
-  const userQuery = query(collection(db, "users"), where("_uid", "==", user.uid));
+  const userQuery = query(collection(db, 'users'), where('_uid', '==', user.uid));
   const querySnapshot = await getDocs(userQuery);
 
   if (querySnapshot.empty) {
-    return res.status(400).send("User not found");
+    return res.status(400).send('User not found');
   } else {
     const userData = querySnapshot.docs[0].data();
     return res.status(200).send(userData);
@@ -127,22 +125,22 @@ const loginUser = async (req, res) => {
   try {
     const { email, password } = req.body;
     if (!email || !password) {
-      return res.status(400).send("Missing email or password");
+      return res.status(400).send('Missing email or password');
     }
 
     const user = await signInWithEmailAndPassword(auth, email, password);
     if (!user) {
-      return res.status(400).send("Invalid email or password");
+      return res.status(400).send('Invalid email or password');
     }
 
     const userObj = user.user;
     const { accessToken, refreshToken } = generateTokens(userObj.uid);
 
-    const userQuery = query(collection(db, "users"), where("_uid", "==", userObj.uid));
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', userObj.uid));
     const querySnapshot = await getDocs(userQuery);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
+      return res.status(400).send('User not found');
     }
 
     const userDocRef = querySnapshot.docs[0].ref;
@@ -164,7 +162,7 @@ const updateFileDetails = async (req, res) => {
   const { uri, newName } = req.body;
 
   if (!uri || !newName) {
-    return res.status(400).send("Missing URI or new name");
+    return res.status(400).send('Missing URI or new name');
   }
 
   try {
@@ -178,28 +176,48 @@ const updateFileDetails = async (req, res) => {
 
 // Delete a file
 const deleteFile = async (req, res) => {
-  const file_url = req.body.file_url;
-  const fileName = file_url.split('/').pop();
+  const fileUrl = req.body.fileName;
   const user = req.body.user.uid;
 
+  if (!fileUrl) {
+    return res.status(400).send('File URL is required');
+  }
+
   try {
-    const userQuery = query(collection(db, "users"), where("_uid", "==", user));
+    // Extract the file name from the URL
+    const fileName = fileUrl.split('/').pop();
+
+    // Query the user's documents
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', user));
     const querySnapshot = await getDocs(userQuery);
-    const blob = bucket.file(fileName);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
-    } else {
-      const userDocRef = querySnapshot.docs[0].ref;
-      await updateDoc(userDocRef, {
-        posts: arrayRemove(file_url)
-      });
+      return res.status(400).send('User not found');
     }
 
+    const userDocRef = querySnapshot.docs[0].ref;
+    const userDocData = querySnapshot.docs[0].data();
+
+    // Check if the file exists in the user's posts
+    const fileExists = userDocData.posts.some(post => post.url === fileUrl);
+
+    if (!fileExists) {
+      return res.status(404).send('File not found in user\'s documents');
+    }
+
+    // Update the user's Firestore document to remove the file entry
+    await updateDoc(userDocRef, {
+      posts: arrayRemove(userDocData.posts.find(post => post.url === fileUrl))
+    });
+
+    // Delete the file from Firebase Storage
+    const blob = bucket.file(fileName);
     await blob.delete();
-    res.status(200).send('File deleted');
+
+    res.status(200).send('File deleted successfully');
   } catch (error) {
-    res.status(400).send('Error deleting file');
+    console.error(error);
+    res.status(500).send('Error deleting file');
   }
 };
 
@@ -230,11 +248,11 @@ const uploadFile = async (req, res) => {
     const publicUrl = `https://storage.googleapis.com/${bucket.name}/${user}_${req.file.originalname}`;
     await blob.makePublic();
 
-    const userQuery = query(collection(db, "users"), where("_uid", "==", user));
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', user));
     const querySnapshot = await getDocs(userQuery);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
+      return res.status(400).send('User not found');
     } else {
       const userDocRef = querySnapshot.docs[0].ref;
       await updateDoc(userDocRef, {
@@ -253,70 +271,104 @@ const uploadFile = async (req, res) => {
 
   blobStream.end(req.file.buffer);
 };
-
-// Rename a file
 const renameFile = async (req, res) => {
   try {
     const user = req.body.user.uid;
-    const oldFileName = req.params.fileName;
-    let newFileName = req.body.newFileName;
+    const oldFileName = `${req.params.fileName}`;
+    const newFileNameWithoutExt = req.body.newFileName;
 
-    if (!oldFileName || !newFileName) {
-      return res.status(400).send("Missing old or new file name");
+    if (!oldFileName || !newFileNameWithoutExt) {
+      return res.status(400).send('Missing old or new file name');
     }
 
-    const userQuery = query(collection(db, "users"), where("_uid", "==", user));
+    // Extract the file extension from the old file name
+    const oldFileExtension = oldFileName.split('.').pop();
+    
+    // Create the new file name with the correct extension
+    const newFileName = `${newFileNameWithoutExt}.${oldFileExtension}`;
+
+    console.log(`Attempting to rename file from: ${oldFileName} to: ${newFileName}`);
+
+    // Check if the old file exists
+    const oldBlob = bucket.file(oldFileName);
+    const [exists] = await oldBlob.exists();
+
+    if (!exists) {
+      console.log(`File ${oldFileName} does not exist`);
+      return res.status(404).send('File not found in Firebase Storage');
+    }
+
+    // Check if the user exists in Firestore
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', user));
     const querySnapshot = await getDocs(userQuery);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
+      console.log(`User with UID ${user} not found`);
+      return res.status(400).send('User not found');
     }
 
     const userDocRef = querySnapshot.docs[0].ref;
     const userDoc = querySnapshot.docs[0].data();
 
-    // Update the file reference in Firestore
-    const updatedPosts = userDoc.posts.map(post => 
-      post.url.includes(oldFileName) ? { ...post, url: post.url.replace(oldFileName, newFileName), name: newFileName } : post
-    );
+    // Log current posts
+    console.log('Current posts:', userDoc.posts);
+
+    // Update file details in user's document
+    const updatedPosts = userDoc.posts.map(post => {
+      if (post.url.includes(oldFileName)) {
+        const newUrl = post.url.replace(oldFileName, newFileName);
+        return { ...post, url: newUrl, name: newFileNameWithoutExt };
+      }
+      return post;
+    });
+
+    console.log('Updated posts:', updatedPosts);
 
     await updateDoc(userDocRef, { posts: updatedPosts });
 
-    // Rename the file in Firebase Storage
-    const oldBlob = bucket.file(oldFileName);
+    // Copy the file to the new name and then delete the old file
     const newBlob = bucket.file(newFileName);
-
     await new Promise((resolve, reject) => {
       oldBlob.copy(newBlob, (err) => {
         if (err) {
+          console.error(`Error copying file: ${err.message}`);
           reject(err);
         } else {
-          oldBlob.delete().then(() => resolve()).catch(err => reject(err));
+          oldBlob.delete().then(() => resolve()).catch(err => {
+            console.error(`Error deleting old file: ${err.message}`);
+            reject(err);
+          });
         }
       });
     });
 
-    res.status(200).send('File renamed successfully');
+    res.status(200).json({ message: 'File renamed successfully', newFileName });
   } catch (error) {
-    console.error(error);
+    console.error(`Error renaming file: ${error.message}`);
     res.status(400).send('Error renaming file');
   }
 };
 
+
 // Like a document
 const likeDocument = async (req, res) => {
-  const { userId, fileUrl } = req.body;
+  const userId = req.body.userId;
+  const fileUrl = req.body.fileUrl;
+
+  console.log('reqBodyLike  ', userId, fileUrl);
+
+  
 
   if (!userId || !fileUrl) {
-    return res.status(400).send("Missing user ID or file URL");
+    return res.status(400).send('Missing user ID or file URL');
   }
 
   try {
-    const userQuery = query(collection(db, "users"), where("_uid", "==", userId));
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', userId));
     const querySnapshot = await getDocs(userQuery);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
+      return res.status(400).send('User not found');
     }
 
     const userDocRef = querySnapshot.docs[0].ref;
@@ -336,16 +388,18 @@ const likeDocument = async (req, res) => {
 const unlikeDocument = async (req, res) => {
   const { userId, fileUrl } = req.body;
 
+  console.log('reqBody', userId, fileUrl);
+
   if (!userId || !fileUrl) {
-    return res.status(400).send("Missing user ID or file URL");
+    return res.status(400).send('Missing user ID or file URL');
   }
 
   try {
-    const userQuery = query(collection(db, "users"), where("_uid", "==", userId));
+    const userQuery = query(collection(db, 'users'), where('_uid', '==', userId));
     const querySnapshot = await getDocs(userQuery);
 
     if (querySnapshot.empty) {
-      return res.status(400).send("User not found");
+      return res.status(400).send('User not found');
     }
 
     const userDocRef = querySnapshot.docs[0].ref;
@@ -360,15 +414,14 @@ const unlikeDocument = async (req, res) => {
     res.status(400).send('Error unliking document');
   }
 };
-
 // Export the routes
 router.post('/register', registerUser);
 router.post('/login', loginUser);
-router.get('/user', authMiddleware, getUser);
+router.get('/', authMiddleware, getUser);
 router.put('/updateFileDetails', authMiddleware, updateFileDetails);
 router.delete('/deleteFile', authMiddleware, deleteFile);
-router.post('/upload', authMiddleware, upload.single('file'), uploadFile);
-router.put('/renameFile/:fileName', authMiddleware, renameFile);
+router.post('/upload', upload.single('file'), authMiddleware, uploadFile);
+router.patch('/renameFile/:fileName', authMiddleware, renameFile);
 router.post('/like', authMiddleware, likeDocument); // New route for liking documents
 router.post('/unlike', authMiddleware, unlikeDocument); // New route for unliking documents
